Move test helpers to test source folder
