+++ java/ingest/ingest-core/src/test/java/sleeper/ingest/job/status/TestIngestJobStatus.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2022 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package sleeper.ingest.job.status;
+
+import sleeper.core.record.process.status.ProcessRun;
+import sleeper.core.record.process.status.ProcessStartedStatus;
+import sleeper.ingest.job.IngestJob;
+
+import java.time.Instant;
+
+public class TestIngestJobStatus {
+
+    public static IngestJobStatus started(IngestJob ingestJob, String taskId, Instant updateTime, Instant startTime) {
+        return from(ingestJob)
+                .jobRun(ProcessRun.started(taskId,
+                        ProcessStartedStatus.updateAndStartTime(updateTime, startTime)))
+                .build();
+    }
+
+    public static IngestJobStatus.Builder from(IngestJob job) {
+        return IngestJobStatus.builder()
+                .jobId(job.getId())
+                .inputFileCount(job.getFiles().size());
+    }
+}
